feat(headless): create initial new Button API
