+++ packages/headless-styles/src/chakra.ts
@@ -2,7 +2,7 @@ import { ChakraAlert } from './components/Alert/chakraAlert'
 import { ChakraAlertDialog } from './components/AlertDialog/chakraAlertDialog'
 import { ChakraAvatar } from './components/Avatar/chakraAvatar'
 import { ChakraBadge } from './components/Badge/chakraBadge'
-import { ChakraButton } from './components/Button/chakraButton'
+import { ChakraButton } from './components/Button_DEPRECATED/chakraButton'
 import { ChakraCheckbox } from './components/Checkbox/chakraCheckbox'
 import { ChakraErrorMessage } from './components/ErrorMessage/errorMessageJS'
 import { ChakraFieldMessage } from './components/FieldMessage/fieldMessageJS'
