+++ packages/headless-styles/src/components/Button/buttonCSS.module.css
@@ -1,15 +1,13 @@
-.base {
+.btnBase {
   align-items: center;
   appearance: none;
   border: none;
   border-radius: 6px;
   cursor: pointer;
   display: inline-flex;
-  font-family: 'PS TT Commons Roman', 'Gotham SSm A', 'Gotham SSm B', Arial,
-    sans-serif;
-  font-size: 1rem;
-  font-variation-settings: 'wght' 600;
-  font-weight: 600;
+  font-family: inherit;
+  font-size: inherit;
+  font-weight: inherit;
   gap: 0.5rem;
   height: 2.5rem;
   justify-content: center;
@@ -26,161 +24,56 @@
   white-space: nowrap;
 }
 
-.text {
-  composes: base;
-  background-color: transparent;
-  color: var(--ps-neutral-text-weak);
+/* sentiment */
+
+.actionButton {
+  composes: btnBase;
 }
 
-.textWeak {
-  composes: base;
-  background-color: transparent;
-  color: var(--ps-text-medium);
+.defaultButton {
+  composes: btnBase;
 }
 
-.weak {
-  composes: base;
-  background-color: var(--ps-neutral-background);
-  color: var(--ps-neutral-text);
+.dangerButton {
+  composes: btnBase;
 }
 
-.medium {
-  composes: base;
-  background-color: var(--ps-background);
-  color: #fff;
+/* usage */
+
+.filledButton {
 }
 
-.strong {
-  composes: base;
-  background-color: var(--ps-background-weak);
-  color: var(--ps-text);
+.outlineButton {
 }
 
-.xs {
-  composes: base;
-  font-size: 0.75rem;
-  gap: 0.25rem;
-  height: 1.5rem;
-  min-width: 1.5rem;
-  padding-inline: 8px;
+.textButton {
 }
 
-.s {
-  composes: base;
-  font-size: 0.875rem;
-  gap: 0.375rem;
-  height: 2rem;
-  min-width: 2rem;
-  padding-inline: 12px;
+/* size */
+
+.mButton {
 }
 
-.m {
-  composes: base;
+.lButton {
 }
 
-.l {
-  composes: base;
-  height: 3rem;
-  min-width: 3rem;
-  padding-inline: 24px;
+/* psuedo styles */
+
+.btnBase:hover {
 }
 
-.textDanger {
-  composes: base;
-  background-color: transparent;
-  color: var(--ps-danger-text-weak);
+.btnBase:disabled {
 }
 
-.mediumDanger {
-  composes: base;
-  background-color: var(--ps-danger-background);
-  color: hsl(336deg 100% 97% / 100%);
+.btnBase:active {
 }
 
-.strongDanger {
-  composes: base;
-  background-color: var(--ps-danger-background-weak);
-  color: var(--ps-danger-text);
+.btnBase:focus {
+  outline: 3px solid var(--ps-info-border);
+  outline-offset: 2px;
 }
 
-.baseIconButton {
-  padding-inline: 0;
-}
-
-.defaultIconButton {
-  composes: baseIconButton;
-}
-
-.roundIconButton {
-  composes: baseIconButton;
-  border-radius: 50%;
-}
-
-/* hover styles */
-
-.base:hover {
-  color: #fff;
-}
-
-.textWeak:hover,
-.medium:hover,
-.strong:hover {
-  background-color: var(--ps-background-hover);
-}
-
-.text:hover,
-.weak:hover {
-  background-color: var(--ps-neutral-background-hover);
-}
-
-.textDanger:hover,
-.mediumDanger:hover,
-.strongDanger:hover {
-  background-color: var(--ps-danger-background-hover);
-}
-
-/* active styles */
-
-.base:active {
-  outline: none;
-}
-
-.text:active,
-.textWeak:active,
-.medium:active,
-.strong:active {
-  background-color: var(--ps-background-active);
-}
-
-.weak:active {
-  background-color: var(--ps-neutral-background-active);
-}
-
-.textDanger:active,
-.mediumDanger:active,
-.strongDanger:active {
-  background-color: var(--ps-danger-background-active);
-}
-
-/* focus styles */
-
-.base:focus {
-  outline: 3px solid var(--ps-background-active);
-}
-
-.base:focus:not(:focus-visible) {
+.btnBase:focus:not(:focus-visible) {
   box-shadow: none;
   outline: none;
-}
-
-/* light theme */
-
-.light .base.text:hover,
-html[data-theme='light'] .base.text:hover {
-  color: var(--ps-neutral-text-weak);
-}
-
-.light .weak:hover,
-html[data-theme='light'] .base.weak:hover {
-  color: var(--ps-neutral-weak);
 }
