+++ packages/headless-styles/src/components/Button_DEPRECATED/buttonCSS.module.css
@@ -1,0 +1,186 @@
+.base {
+  align-items: center;
+  appearance: none;
+  border: none;
+  border-radius: 6px;
+  cursor: pointer;
+  display: inline-flex;
+  font-family: 'PS TT Commons Roman', 'Gotham SSm A', 'Gotham SSm B', Arial,
+    sans-serif;
+  font-size: 1rem;
+  font-variation-settings: 'wght' 600;
+  font-weight: 600;
+  gap: 0.5rem;
+  height: 2.5rem;
+  justify-content: center;
+  min-width: 2.5rem;
+  outline: none;
+  padding-inline: 16px;
+  position: relative;
+  text-align: center;
+  text-decoration: none;
+  text-transform: none;
+  transition: background-color 250ms ease-in-out, color 250ms ease-in-out;
+  user-select: none;
+  vertical-align: middle;
+  white-space: nowrap;
+}
+
+.text {
+  composes: base;
+  background-color: transparent;
+  color: var(--ps-neutral-text-weak);
+}
+
+.textWeak {
+  composes: base;
+  background-color: transparent;
+  color: var(--ps-text-medium);
+}
+
+.weak {
+  composes: base;
+  background-color: var(--ps-neutral-background);
+  color: var(--ps-neutral-text);
+}
+
+.medium {
+  composes: base;
+  background-color: var(--ps-background);
+  color: #fff;
+}
+
+.strong {
+  composes: base;
+  background-color: var(--ps-background-weak);
+  color: var(--ps-text);
+}
+
+.xs {
+  composes: base;
+  font-size: 0.75rem;
+  gap: 0.25rem;
+  height: 1.5rem;
+  min-width: 1.5rem;
+  padding-inline: 8px;
+}
+
+.s {
+  composes: base;
+  font-size: 0.875rem;
+  gap: 0.375rem;
+  height: 2rem;
+  min-width: 2rem;
+  padding-inline: 12px;
+}
+
+.m {
+  composes: base;
+}
+
+.l {
+  composes: base;
+  height: 3rem;
+  min-width: 3rem;
+  padding-inline: 24px;
+}
+
+.textDanger {
+  composes: base;
+  background-color: transparent;
+  color: var(--ps-danger-text-weak);
+}
+
+.mediumDanger {
+  composes: base;
+  background-color: var(--ps-danger-background);
+  color: hsl(336deg 100% 97% / 100%);
+}
+
+.strongDanger {
+  composes: base;
+  background-color: var(--ps-danger-background-weak);
+  color: var(--ps-danger-text);
+}
+
+.baseIconButton {
+  padding-inline: 0;
+}
+
+.defaultIconButton {
+  composes: baseIconButton;
+}
+
+.roundIconButton {
+  composes: baseIconButton;
+  border-radius: 50%;
+}
+
+/* hover styles */
+
+.base:hover {
+  color: #fff;
+}
+
+.textWeak:hover,
+.medium:hover,
+.strong:hover {
+  background-color: var(--ps-background-hover);
+}
+
+.text:hover,
+.weak:hover {
+  background-color: var(--ps-neutral-background-hover);
+}
+
+.textDanger:hover,
+.mediumDanger:hover,
+.strongDanger:hover {
+  background-color: var(--ps-danger-background-hover);
+}
+
+/* active styles */
+
+.base:active {
+  outline: none;
+}
+
+.text:active,
+.textWeak:active,
+.medium:active,
+.strong:active {
+  background-color: var(--ps-background-active);
+}
+
+.weak:active {
+  background-color: var(--ps-neutral-background-active);
+}
+
+.textDanger:active,
+.mediumDanger:active,
+.strongDanger:active {
+  background-color: var(--ps-danger-background-active);
+}
+
+/* focus styles */
+
+.base:focus {
+  outline: 3px solid var(--ps-background-active);
+}
+
+.base:focus:not(:focus-visible) {
+  box-shadow: none;
+  outline: none;
+}
+
+/* light theme */
+
+.light .base.text:hover,
+html[data-theme='light'] .base.text:hover {
+  color: var(--ps-neutral-text-weak);
+}
+
+.light .weak:hover,
+html[data-theme='light'] .base.weak:hover {
+  color: var(--ps-neutral-weak);
+}
